First commit of the project
As-is, it seems to work, but I lack fulltext inputs to test.
I may have luck with gutemberg project in the future, but I need to sort the languages of the books (and find a way to download the huge amount of files).
Found some examples on textfiles.com, but they lack an uniform layout and tend to produce strange result.

Todo:
- Change the regular expression to remove useless symbols.
- Use command-line args.
- Serialize the result of parsing.